apps/preprocessing_gateway/requirements.txt
luigi>=3.5.0
numpy==2.0.0
pandas==2.2.2
python-dotenv==1.0.1
scikit-learn==1.5.1
scipy==1.14.0